requirements.txt
discord.py
rotten_tomatoes_client
